requirements.txt
fastapi
uvicorn
faiss-cpu
numpy
pydantic
openai
faiss-cpu
numpy
pydantic
markitdown
mammoth